+++ examples/basic/credentials.sample.h
@@ -4,4 +4,4 @@
 
 // Check that your server actually returns something in this URL
 // In particular, remember to add a trailing slash for root requests
-#define NOFUSS_SERVER   "http://192.168.1.100/"
+#define FIRMWARE_URL   "http://192.168.1.100/"